Version 2.2.2 (#390)
* Adding an option to use /dev/shm for mikado serialise and compare.

* Bumped version

* Using TMPDIR by default when creating/reading sqlite databases

+++ Mikado/picking/picker.py
@@ -17,12 +17,14 @@ import zlib
 from logging import handlers as logging_handlers
 import functools
 import multiprocessing
+
+from Mikado import version
 from sqlalchemy.engine import create_engine  # SQLAlchemy/DB imports
 from sqlalchemy.orm.session import sessionmaker
 from sqlalchemy.pool import QueuePool as SqlPool
 import sqlalchemy
 import sqlalchemy.exc
-from ..utilities import path_join
+from ..utilities import path_join, create_shm_handle
 from ..utilities.log_utils import formatter
 from ..parsers.GTF import GTF, GtfLine
 from ..parsers.GFF import GFF3
@@ -221,28 +223,19 @@ class Picker:
         This method will copy the SQLite input DB into memory.
         """
 
-        if self.configuration.pick.run_options.shm is True:
-            shm_available = os.path.exists("/dev/shm") and os.access("/dev/shm", os.W_OK)
-            if shm_available is False:
-                self.main_logger.info("Mikado was asked to copy the database into /dev/shm, but it is either \
-not available or not writable for this user. Leaving the DB where it is.")
-                return
-            self.main_logger.info("Copying Mikado database into a SHM db")
-            assert self.configuration.db_settings.dbtype == "sqlite"
-            # Create temporary file
-            self.shm_db = tempfile.NamedTemporaryFile(suffix=".db", prefix="/dev/shm/")
-            self.main_logger.debug("Copying {0} into {1}".format(
-                self.configuration.db_settings.db,
-                self.shm_db.name))
-            try:
-                shutil.copy2(self.configuration.db_settings.db, self.shm_db.name)
-                self.configuration.db_settings.db = self.shm_db.name
-            except PermissionError:
-                self.main_logger.warning(
-                    """Permission to write on /dev/shm denied.
-                    Back to using the DB on disk.""")
-                self.configuration.pick.run_options.shm = False
-            self.main_logger.info("DB copied into memory")
+        if self.configuration.db_settings.dbtype == "sqlite" and os.path.exists(self.configuration.db_settings.db):
+            if self.configuration.pick.run_options.shm is True:
+                handle = create_shm_handle()
+                if handle is None:
+                    self.main_logger.info("Mikado was asked to copy the database into /dev/shm, but it is either "
+                                          "not available or not writable for this user.")
+                    handle = tempfile.NamedTemporaryFile(suffix=".db")
+            else:
+                handle = tempfile.NamedTemporaryFile(suffix=".db")
+            self.shm_db = handle
+            shutil.copy2(self.configuration.db_settings.db, self.shm_db.name)
+            self.configuration.db_settings.db = self.shm_db.name
+            self.main_logger.info(f"DB copied into {self.shm_db.name}")
 
     def setup_logger(self):
 
@@ -299,13 +292,14 @@ not available or not writable for this user. Leaving the DB where it is.")
             self.main_logger.setLevel(logging.INFO)
         self.main_logger.addHandler(self.log_handler)
 
-        self.main_logger.info("Begun analysis of {0}".format(self.input_file))
+        self.main_logger.info(f"Mikado version: {version.__version__}")
         if self.commandline != '':
             self.main_logger.info("Command line: {0}".format(self.commandline))
         else:
             self.main_logger.info(
                 "Analysis launched directly, without using the launch script.")
 
+        self.main_logger.info("Begun analysis of {0}".format(self.input_file))
         # Create the shared DB if necessary
         self.log_writer = logging_handlers.QueueListener(self.logging_queue, self.logger)
         self.log_writer.start()
